refactor(Chat): split Chat component into modular components and optimize performance
- Decompose monolithic Chat component into smaller focused components:
 - ChatContainer: Top-level socket management
 - ChatList: Message display and scroll management
 - ChatInput: Message input and submission
 - ChatBubble: Individual message UI
- Add memo to all components to prevent unnecessary re-renders
- Implement individual selectors for Zustand store access
- Enhance useDrawingSocket hook with submitDrawing event handler
- Improve code organization and separation of concerns
